% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.plot.R
\name{quantile.rug}
\alias{quantile.rug}
\title{Quantile rug plot}
\usage{
\method{quantile}{rug}(x, prob = 0:10/10, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{prob}{The quantiles of x to mark on the x-axis.}

\item{...}{Additional optional arguments to be passed onto 
\code{\link[graphics]{rug}}}
}
\value{
No return values.
}
\description{
Marks the quantiles on the axes of the current plot.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
plot(x, y)
quantile.rug(x)
}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[stats]{quantile}},
\code{\link[base]{jitter}}, \code{\link[graphics]{rug}}.
}
\author{
Greg Ridgeway \email{gregridgeway@gmail.com}.
}
\keyword{aplot}
