% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gbm.R
\name{test.gbm}
\alias{test.gbm}
\alias{validate.gbm}
\alias{test.relative.influence}
\title{Test the \code{gbm} package.}
\usage{
test.gbm()
}
\value{
An object of class \code{RUnitTestData}. See the help for
\code{RUnit} for details.
}
\description{
Run tests on \code{gbm} functions to perform logical checks and
reproducibility.
}
\details{
The function uses functionality in the \code{RUnit} package. A fairly small
validation suite is executed that checks to see that relative influence
identifies sensible variables from simulated data, and that predictions from
GBMs with Gaussian, Cox or binomial distributions are sensible,
}
\note{
The test suite is not comprehensive.
}
\examples{

# Uncomment the following lines to run - commented out to make CRAN happy
#library(RUnit)
#val <- validate.texmex()
#printHTMLProtocol(val, "texmexReport.html")
}
\seealso{
\code{\link{gbm}}
}
\author{
Harry Southworth
}
\keyword{models}
