% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm-package.R
\docType{package}
\name{gbm-package}
\alias{gbm-package}
\title{Generalized Boosted Regression Models (GBMs)}
\description{
This package implements extensions to Freund and Schapire's AdaBoost
algorithm and J. Friedman's gradient boosting machine. Includes regression
methods for least squares, absolute loss, logistic, Poisson, Cox
proportional hazards partial likelihood, multinomial, t-distribution,
AdaBoost exponential loss, Learning to Rank, and Huberized hinge loss.
This gbm package is no longer under further development. Consider
https://github.com/gbm-developers/gbm3 for the latest version.
}
\details{
Further information is available in vignette: 
\code{browseVignettes(package = "gbm")}
}
\references{
Y. Freund and R.E. Schapire (1997) \dQuote{A decision-theoretic
generalization of on-line learning and an application to boosting,}
\emph{Journal of Computer and System Sciences,} 55(1):119-139.

G. Ridgeway (1999). \dQuote{The state of boosting,} \emph{Computing Science
and Statistics} 31:172-181.

J.H. Friedman, T. Hastie, R. Tibshirani (2000). \dQuote{Additive Logistic
Regression: a Statistical View of Boosting,} \emph{Annals of Statistics}
28(2):337-374.

J.H. Friedman (2001). \dQuote{Greedy Function Approximation: A Gradient
Boosting Machine,} \emph{Annals of Statistics} 29(5):1189-1232.

J.H. Friedman (2002). \dQuote{Stochastic Gradient Boosting,}
\emph{Computational Statistics and Data Analysis} 38(4):367-378.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gbm-developers.github.io/gbm/}
  \item \url{https://github.com/gbm-developers/gbm}
  \item Report bugs at \url{https://github.com/gbm-developers/gbm/issues}
}

}
\author{
Greg Ridgeway \email{gridge@upenn.edu} with contributions by
Daniel Edwards, Brian Kriegler, Stefan Schroedl, Harry Southworth,
and Brandon Greenwell
}
\keyword{package}
